requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
google-genai>=0.5.0
python-dotenv>=1.0.0

# Tunnel providers
pyngrok>=7.0.0

# Utilities
colorama>=0.4.6
requests>=2.31.0

# Utilities
colorama>=0.4.6
requests>=2.31.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
